-[ ] updates Navbar -[ ] create file Sidebar -[ ] create file icon and page in constant folder
