--- src/Components/Button.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Button = () => {
-  return (
-    <div>
-      <button>button</button>
-    </div>
-  );
-};
-
-export default Button;
