--- src/app/types/icons.tsx
@@ -1,6 +1,0 @@
-// export default interface KochbhiTypes {
-
-//     id:string;
-//     skip: number;
-
-// }
