+++ src/constant/icons.tsx
@@ -1,0 +1,11 @@
+import Heart from '@/../public/icons/heart.svg';
+import Profile from '@/../public/icons/profile.svg';
+import Shop from '@/../public/icons/shop.svg';
+import Search from '@/../public/icons/search.svg';
+
+export const icons = [
+  { icon: Search.src, link: '/search' },
+  { icon: Heart.src, link: '/favorite' },
+  { icon: Shop.src, link: '/shop' },
+  { icon: Profile.src, link: '/profile' },
+];
