+++ src/constant/pagelist.ts
@@ -1,0 +1,6 @@
+export const pages = [
+  { text: 'Home', link: '/' },
+  { text: 'Product', link: '/product' },
+  { text: 'About Us', link: '/about' },
+  { text: 'Contact Us', link: '/contact' },
+];
